Разделение кода на модули. Редактирование кода.
