+++ js/string-length.js
@@ -1,0 +1,6 @@
+// Функция для проверки максимальной длины строки.
+const stringLength = function isLengthLongCorrect (string, maxStringLength) {
+  return string.length >= maxStringLength;
+};
+
+export {stringLength};
